+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.0.8
 - Whether player level bonuses bypass the max level cap is now configurable
+- Entity blacklist/whitelists now support modid wildcards (`cataclysm:*`) (This was already supported, but undocumented)
 
 ## 1.0.7
 - Change default `#c:is_cold/overworld` biome level bonus to `#c:is_icy`